# api/openapi.yaml
openapi: 3.1.0
info:
  title: Collections Agent API
  version: 1.0.0
  description: REST API for the AI Receivables Collections Agent console, the bank webhook and admin controls. Serves the
    stories named in each operation's x-story-ids. Money fields are integer paise named *_paise (INR only); every amount shown
    to a user is formatted from these at the edge.
  contact:
    name: HACK-001 team
...
servers:
- url: http://localhost:8000/api/v1
  description: local Docker Compose
security:
- demoRole: []
...
paths:
...
  /auth/me:
    get:
      operationId: getMe
      summary: Current user
      tags:
      - auth
      x-story-ids:
      - US-01-007
      x-roles: viewer, collector, admin
      parameters:
      - $ref: '#/components/parameters/ApiVersion'
      responses:
        '200':
          description: OK
          headers:
            X-Request-Id:
              $ref: '#/components/headers/XRequestId'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '500':
          $ref: '#/components/responses/InternalError'
  /dashboard:
    get:
      operationId: getDashboard
      summary: Totals, ageing and the attention list
      tags:
      - dashboard
      x-story-ids:
      - US-00-003
      x-roles: viewer, collector, admin
      parameters:
      - $ref: '#/components/parameters/ApiVersion'
      responses:
        '200':
          description: OK
          headers:
            X-Request-Id:
              $ref: '#/components/headers/XRequestId'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Dashboard'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '500':
          $ref: '#/components/responses/InternalError'
  /executive:
    get:
      operationId: getExecutive
      summary: "CFO dashboard: KPIs, series and the attention list with factors"
      description: Every figure is computed from the ledger; the definitions field states each one. Shape in backend/app/services/executive.py.
      tags:
      - dashboard
      x-story-ids:
      - HACK-003
      x-roles: viewer, collector, admin
      parameters:
      - $ref: '#/components/parameters/ApiVersion'
      responses:
        '200':
          description: OK
          headers:
            X-Request-Id:
              $ref: '#/components/headers/XRequestId'
          content:
            application/json:
              schema:
                type: object
        '401':
          $ref: '#/components/responses/Unauthorized'
        '500':
          $ref: '#/components/responses/InternalError'
  /safety:
    get:
      operationId: getSafety
      summary: "AI Safety Center: counts of guardrail, approval and send events"
      description: Counts of real rows only; the sources field states what each count is made of. Shape in backend/app/services/safety.py.
      tags:
      - dashboard
      x-story-ids:
      - HACK-003
      x-roles: viewer, collector, admin
      parameters:
      - $ref: '#/components/parameters/ApiVersion'
      responses:
        '200':
          description: OK
          headers:
            X-Request-Id:
              $ref: '#/components/headers/XRequestId'
          content:
            application/json:
              schema:
                type: object
        '401':
          $ref: '#/components/responses/Unauthorized'
        '500':
          $ref: '#/components/responses/InternalError'
...
  /customers/{id}/timeline:
    get:
      operationId: listCustomerTimeline
      summary: Timeline, newest first, and the next action
      tags:
      - customers
      x-story-ids:
      - US-00-022
      - US-00-004
      x-roles: viewer, collector, admin
      parameters:
      - $ref: '#/components/parameters/ApiVersion'
      - name: id
        in: path
        required: true
        schema:
          type: string
          format: uuid
      responses:
        '200':
          description: OK
          headers:
            X-Request-Id:
              $ref: '#/components/headers/XRequestId'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Timeline'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalError'
  /customers/{id}/runs:
    get:
      operationId: listCustomerRuns
      summary: Agent runs for the customer
      tags:
      - runs
      x-story-ids:
      - US-00-008
      x-roles: viewer, collector, admin
      parameters:
      - $ref: '#/components/parameters/ApiVersion'
      - $ref: '#/components/parameters/Cursor'
      - $ref: '#/components/parameters/Limit'
      - name: id
        in: path
        required: true
        schema:
          type: string
          format: uuid
      responses:
        '200':
          description: OK
          headers:
            X-Request-Id:
              $ref: '#/components/headers/XRequestId'
          content:
            application/json:
              schema:
                type: object
                required:
                - data
                - page
                properties:
                  data:
                    type: array
                    items:
                      $ref: '#/components/schemas/AgentRun'
                  page:
                    $ref: '#/components/schemas/Page'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalError'
  /customers/{id}/call-prep:
    get:
      operationId: getCallPrep
      summary: Prepare Call sheet (FEATURE_VOICE)
      tags:
      - customers
      x-story-ids:
      - US-00-025
      x-roles: collector, admin
      parameters:
      - $ref: '#/components/parameters/ApiVersion'
      - name: id
        in: path
        required: true
        schema:
          type: string
          format: uuid
      responses:
        '200':
          description: OK
          headers:
            X-Request-Id:
              $ref: '#/components/headers/XRequestId'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CallPrep'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalError'
...
  /promises:
    get:
      operationId: listPromises
      summary: 'Promises (today''s: filter[promised_date]=today)'
      tags:
      - promises
      x-story-ids:
      - US-00-021
      - US-00-020
      x-roles: viewer, collector, admin
      parameters:
      - $ref: '#/components/parameters/ApiVersion'
      - $ref: '#/components/parameters/Cursor'
      - $ref: '#/components/parameters/Limit'
      - name: filter[status]
        in: query
        required: false
        schema:
          type: string
          enum:
          - pending
          - fulfilled
          - partially_fulfilled
          - missed
      - name: filter[promised_date]
        in: query
        required: false
        schema:
          type: string
          format: date
      responses:
        '200':
          description: OK
          headers:
            X-Request-Id:
              $ref: '#/components/headers/XRequestId'
          content:
            application/json:
              schema:
                type: object
                required:
                - data
                - page
                properties:
                  data:
                    type: array
                    items:
                      $ref: '#/components/schemas/Promise'
                  page:
                    $ref: '#/components/schemas/Page'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '500':
          $ref: '#/components/responses/InternalError'
  /promises/{id}/check-payment:
    post:
      operationId: checkPromisePayment
      summary: Check payment for one promise
      tags:
      - promises
      x-story-ids:
      - US-00-021
      x-roles: collector, admin
      parameters:
      - $ref: '#/components/parameters/ApiVersion'
      - name: id
        in: path
        required: true
        schema:
          type: string
          format: uuid
      responses:
        '200':
          description: OK
          headers:
            X-Request-Id:
              $ref: '#/components/headers/XRequestId'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PromiseCheck'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalError'
  /disputes:
    get:
      operationId: listDisputes
      summary: Disputes
      tags:
      - disputes
      x-story-ids:
      - US-00-016
      x-roles: viewer, collector, admin
      parameters:
      - $ref: '#/components/parameters/ApiVersion'
      - $ref: '#/components/parameters/Cursor'
      - $ref: '#/components/parameters/Limit'
      - name: filter[status]
        in: query
        required: false
        schema:
          type: string
          enum:
          - open
          - resolved
      responses:
        '200':
          description: OK
          headers:
            X-Request-Id:
              $ref: '#/components/headers/XRequestId'
          content:
            application/json:
              schema:
                type: object
                required:
                - data
                - page
                properties:
                  data:
                    type: array
                    items:
                      $ref: '#/components/schemas/Dispute'
                  page:
                    $ref: '#/components/schemas/Page'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '500':
          $ref: '#/components/responses/InternalError'
  /disputes/{id}/resolve:
    post:
      operationId: resolveDispute
      summary: Resolve a dispute; reminders resume
      tags:
      - disputes
      x-story-ids:
      - US-00-016
      x-roles: collector, admin
      parameters:
      - $ref: '#/components/parameters/ApiVersion'
      - name: id
        in: path
        required: true
        schema:
          type: string
          format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Resolve'
      responses:
        '200':
          description: OK
          headers:
            X-Request-Id:
              $ref: '#/components/headers/XRequestId'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Dispute'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '409':
          $ref: '#/components/responses/Conflict'
        '422':
          $ref: '#/components/responses/UnprocessableEntity'
        '500':
          $ref: '#/components/responses/InternalError'
  /escalations:
    get:
      operationId: listEscalations
      summary: Escalations
      tags:
      - escalations
      x-story-ids:
      - US-00-016
      - US-00-012
      x-roles: viewer, collector, admin
      parameters:
      - $ref: '#/components/parameters/ApiVersion'
      - $ref: '#/components/parameters/Cursor'
      - $ref: '#/components/parameters/Limit'
      - name: filter[status]
        in: query
        required: false
        schema:
          type: string
          enum:
          - open
          - resolved
      responses:
        '200':
          description: OK
          headers:
            X-Request-Id:
              $ref: '#/components/headers/XRequestId'
          content:
            application/json:
              schema:
                type: object
                required:
                - data
                - page
                properties:
                  data:
                    type: array
                    items:
                      $ref: '#/components/schemas/Escalation'
                  page:
                    $ref: '#/components/schemas/Page'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '500':
          $ref: '#/components/responses/InternalError'
  /escalations/{id}/resolve:
    post:
      operationId: resolveEscalation
      summary: Close an escalation
      tags:
      - escalations
      x-story-ids:
      - US-00-016
      x-roles: collector, admin
      parameters:
      - $ref: '#/components/parameters/ApiVersion'
      - name: id
        in: path
        required: true
        schema:
          type: string
          format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Resolve'
      responses:
        '200':
          description: OK
          headers:
            X-Request-Id:
              $ref: '#/components/headers/XRequestId'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Escalation'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '409':
          $ref: '#/components/responses/Conflict'
        '422':
          $ref: '#/components/responses/UnprocessableEntity'
        '500':
          $ref: '#/components/responses/InternalError'
  /payments:
    get:
      operationId: listPayments
      summary: Payments
      tags:
      - payments
      x-story-ids:
      - US-00-017
      x-roles: viewer, collector, admin
      parameters:
      - $ref: '#/components/parameters/ApiVersion'
      - $ref: '#/components/parameters/Cursor'
      - $ref: '#/components/parameters/Limit'
      - name: filter[match_status]
        in: query
        required: false
        schema:
          type: string
          enum:
          - matched
          - needs_verification
      responses:
        '200':
          description: OK
          headers:
            X-Request-Id:
              $ref: '#/components/headers/XRequestId'
          content:
            application/json:
              schema:
                type: object
                required:
                - data
                - page
                properties:
                  data:
                    type: array
                    items:
                      $ref: '#/components/schemas/Payment'
                  page:
                    $ref: '#/components/schemas/Page'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '500':
          $ref: '#/components/responses/InternalError'
  /payments/{id}/match:
    post:
      operationId: matchPayment
      summary: Collector matches a payment that needs verification
      tags:
      - payments
      x-story-ids:
      - US-00-017
      - US-00-018
      x-roles: collector, admin
      parameters:
      - $ref: '#/components/parameters/ApiVersion'
      - name: id
        in: path
        required: true
        schema:
          type: string
          format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PaymentMatch'
      responses:
        '200':
          description: OK
          headers:
            X-Request-Id:
              $ref: '#/components/headers/XRequestId'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Payment'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '409':
          $ref: '#/components/responses/Conflict'
        '422':
          $ref: '#/components/responses/UnprocessableEntity'
        '500':
          $ref: '#/components/responses/InternalError'
  /webhooks/bank:
    post:
      operationId: receiveBankCredit
      summary: Signed bank credit (HMAC-SHA256 over timestamp.body)
      tags:
      - webhooks
      x-story-ids:
      - US-00-017
      x-roles: bank feed
      parameters:
      - $ref: '#/components/parameters/ApiVersion'
      - $ref: '#/components/parameters/BankTimestamp'
      security:
      - bankSignature: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/BankCredit'
      responses:
        '200':
          description: OK
          headers:
            X-Request-Id:
              $ref: '#/components/headers/XRequestId'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Payment'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '422':
          $ref: '#/components/responses/UnprocessableEntity'
        '500':
          $ref: '#/components/responses/InternalError'
...
  /evals/latest:
    get:
      operationId: getLatestEval
      summary: Latest evaluation report
      tags:
      - evals
      x-story-ids:
      - US-01-010
      x-roles: viewer, collector, admin
      parameters:
      - $ref: '#/components/parameters/ApiVersion'
      responses:
        '200':
          description: OK
          headers:
            X-Request-Id:
              $ref: '#/components/headers/XRequestId'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/EvalReport'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalError'
...
  /invoices/{number}/pay-link:
    post:
      operationId: createPayLink
      summary: Signed SIMULATED payment link for one invoice (FEATURE_PAYMENT_LINK)
      tags:
      - pay
      x-story-ids:
      - US-03-004
      x-roles: collector, admin
      parameters:
      - $ref: '#/components/parameters/ApiVersion'
      - name: number
        in: path
        required: true
        schema:
          type: string
          pattern: ^INV-\d+$
      responses:
        '200':
          description: OK
          headers:
            X-Request-Id:
              $ref: '#/components/headers/XRequestId'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PayLinkToken'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalError'
...
  /follow-ups:
    get:
      operationId: listFollowUps
      summary: Follow-up tasks for missed and partly kept promises
      tags:
      - promises
      x-story-ids:
      - HACK-003
      x-roles: viewer, collector, admin
      parameters:
      - $ref: '#/components/parameters/ApiVersion'
      - name: filter[customer_id]
        in: query
        required: false
        schema:
          type: string
      - name: filter[status]
        in: query
        required: false
        schema:
          type: string
      - name: limit
        in: query
        required: false
        schema:
          type: string
      responses:
        '200':
          description: OK
          headers:
            X-Request-Id:
              $ref: '#/components/headers/XRequestId'
          content:
            application/json:
              schema:
                type: object
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '500':
          $ref: '#/components/responses/InternalError'
...
components:
  securitySchemes:
    demoRole:
      type: apiKey
      in: header
      name: X-Demo-Role
      description: admin, collector or viewer
    bankSignature:
      type: apiKey
      in: header
      name: X-Bank-Signature
      description: hex HMAC-SHA256 of '<X-Bank-Timestamp>.<raw body>' with BANK_WEBHOOK_SECRET
  parameters:
    ApiVersion:
      name: X-API-Version
      in: header
      required: false
      schema:
        type: string
        example: '2026-09-30'
...
    Cursor:
      name: cursor
      in: query
      required: false
      schema:
        type: string
...
  headers:
    XRequestId:
      schema:
        type: string
      description: Matches error.request_id and the server log line
...
  schemas:
...
    User:
      type: object
      properties:
        id:
          type: string
          format: uuid
        email:
          type: string
        display_name:
          type: string
        role:
          type: string
          enum:
          - admin
          - collector
          - viewer
      required:
      - id
      - email
      - display_name
      - role
      example:
        id: 0192b1c2-7f3a-7c4e-9a1b-2c3d4e5f6a7b
        email: collector@example.in
        display_name: Priya (Collector)
        role: collector
...